Robust 'signature parsing'; using network sniffing.
Temporary signature parsing code has been removed in favor of more
reliable networking sniffing method. The player instance is modified to
contain valid streams; the stream that will be playing is sniffed from
the network for the signature and set as active on the download buttons.
This change also allows downloading of lower quality streams.